requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
greenlet==1.1.1
idna==3.2
IMDbPY==2021.4.18
lxml
mysql-connector-python
mysqlclient==2.0.3
omdb
parse-torrent-title==2.3
protobuf==3.17.3
PyMySQL==1.0.2
python-dateutil
requests
six==1.16.0
SQLAlchemy
tmdbsimple
tqdm
transmission-rpc==3.2.7
typing-extensions==3.10.0.0
urllib3==1.26.6
wincertstore==0.2
